updated exception hierarchy, refactored
