+++ executor-service/src/main/java/executor/service/exception/scenario/ScenarioExecutionException.java
@@ -1,0 +1,7 @@
+package executor.service.exception.scenario;
+
+public class ScenarioExecutionException extends RuntimeException {
+    public ScenarioExecutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
