+++ executor-service/src/main/java/executor/service/exception/scenario/SiteNotFoundException.java
@@ -1,0 +1,7 @@
+package executor.service.exception.scenario;
+
+public class SiteNotFoundException extends ScenarioExecutionException {
+    public SiteNotFoundException(String url, Throwable cause) {
+        super("Site by url: " + url + " is not found!", cause);
+    }
+}
