+++ executor-service/src/main/java/executor/service/exception/scenario/step/ClickXPathException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.stepexception;
+package executor.service.exception.scenario.step;
 
 
 public class ClickXPathException extends StepExecutionException {
