--- executor-service/src/main/java/executor/service/exception/stepexception/StepExecutionException.java
@@ -1,7 +1,0 @@
-package executor.service.exception.stepexception;
-
-public class StepExecutionException extends RuntimeException {
-    public StepExecutionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
